==> 1_data_collection.ipynb <==
# Notebook description

#### What is this notebook for? 
This this the first notebook of a serie of five. Its main purpose is to collect the data that will be used for the rest of the project. 


The process of data collection will be divided in multiple processes: 
1. Create a list of pages/account where I found bots' comments 
2. For each, I look at their last 50+ posts to collect the ID of the post 
3. Once done, loop through those thousands of IDs to collect the comments
4. 



collected the IDs of  from multiple pages suseptible to have bots commenting on their posts. I selected those pages manually by making sure they're all bots' targets. Then, I'll loop through each post thanks to it's ID. On instagram, each post url is made as `instagram.com/p/{postid}`

## Script

## Import modules and UDFs

#Personnal module, functions I use often
import src.webscraping as mw
import src.useful as mu


# Better print 
from tqdm import tqdm

# To become Dr Strange 
import time 
from datetime import datetime

# Basic data 
import pandas as pd
import numpy as np

# Store data 
import sqlite3
import json
from flatten_json import flatten

# Move things around locally
import shutil

# Fetch instagram data 
from instaloader import Instaloader, Profile
from instaloader.exceptions import ProfileNotExistsException
import urllib
from splinter import Browser

import os

from coolname import generate_slug

### Setting notebook preferences

pd.set_option("display.max_columns", None)

# Creating SQL database to store all the data for this project
database = "data/main_database.sqlite"
con = sqlite3.connect(database)

## 1. Collect post ids

List a number of instagram accounts that have bots commenting on their posts. Here I'm looping through a page list that is targeted by bots and collect the posts_ids one by one.

# Listing pages targeted by bots
pages = ["nfl", "championsleague", "mercedesamgf1", "ESPN", "bleacherreport", "houseofhighlights", "nba", "worldstar", "grmdaily", "pubity", 
         "meme.ig", "brgridiron", "lakers", "ballislife", "nflonfox", "nflnetwork", "espnnfl", "cbssports", "thecheckdown"]
len(pages)

# Lunch browser
browser = Browser('chrome')

post_per_page = 50
for page in pages:
    browser.visit(f"https://www.instagram.com/{page}/")
    postids = []
    

    while len(postids) < post_per_page:
        # Scroll up and then down each time helps the page to not bug
        browser.execute_script("window.scrollTo(0, 0);")
        time.sleep(0.5)
        browser.execute_script("window.scrollTo(0, document.body.scrollHeight);")
        time.sleep(2)
        
        # Instagram doesn't show in the html the posts it of the post we don't see so I need to slowly scroll down to collect each of them.
        browser.execute_script("window.scrollBy(0, 200);")

        # Change the browser to a beautiful soup object where I can get the posts id
        soup = mw.bsoup(browser)
        for element in soup.find_all("a"):
            link = element.get("href")
            if "/p/" in link: # We can find the posts id by looking into a tags that have an attribute of href.
                postids.append(link.replace("/p/", "").replace("/", ""))

        # Create a df with post_ids and save it in the db
        df_post_ids = pd.DataFrame(set(postids), columns=["post_id"])
        df_post_ids['page'] = page
        df_post_ids.to_sql("post_ids", con, if_exists="append", index=False)

        # Display random timer to do nothing
        for i in list(range(np.random.randint(10, 30)))[::-1]:
            print(i, end="\r")
            time.sleep(1)

## 2. Scrape comments

# Query post_id not in comments table 
query = """
SELECT  
    DISTINCT
    post_id
    , page_name
FROM post_ids 
"""

# Loadind post_ids from db
posts_ids_to_scrape = pd.read_sql_query(query, con)
posts_ids_to_scrape

# Scraping and storing comments from post_ids:
np.random.shuffle(posts_ids_to_scrape)
for post_id in tqdm(posts_ids_to_scrape):
    browser.visit(f"https://www.instagram.com/p/{post_id}/")
    
    # Wait for page to load and get its soup
    while True: 
        soup = mw.bsoup(browser)
        if soup.find("h2", class_="_a9zc") != None: 
            break

    # Get days since posted
    post_posted_time = soup.find("time", class_="_a9ze _a9zf").get("datetime")
    now = datetime.now()
    days_diff = (now - pd.to_datetime(post_posted_time[:-1])).days

    # I only keep what was posted less than a month ago so I don't get too old data
    if days_diff > 31: 
        continue
    
    df_post_comments = pd.DataFrame(columns=["post_id", "page", "legend", "post_posted_time", "username", 
                                             "full_comment_data", "comment", "comment_posted_time", 
                                             "time_since_posted", "comments_likes", "replies",
                                             "time_now"])

    for comment_block in soup.find_all("ul", class_="_a9ym"):
        page = soup.find("h2", class_="_a9zc").text 
        legend = soup.find("div", class_="_a9zs").text 
        time_since_posted = comment_block.find("time", class_="_a9ze _a9zf").text 
        username = comment_block.find("h3", class_="_a9zc").text 
        comment_posted_time = comment_block.find("time", class_="_a9ze _a9zf").get("datetime")
        comments_likes = comment_block.find("button", class_="_a9ze").text 
        comment = comment_block.find("div", class_="_a9zs").text  
        replies = comment_block.find("li", class_="_a9yg").text if comment_block.find("li", class_="_a9yg") != None else ""
        full_comment_data = comment_block.text
        time_now = datetime.now()

        # Add comment values to dataframe
        df_post_comments.loc[len(df_post_comments)] = [post_id, page, legend, post_posted_time, username, 
                                                       full_comment_data, comment, comment_posted_time, 
                                                       time_since_posted, comments_likes, replies,
                                                       time_now]

        
    df_post_comments.to_sql("comments", con, if_exists="append", index=False)
    time.sleep(np.random.randint(40, 60))

### Create cooler name 

df_post_comments = pd.read_sql_query('select distinct username from comments', con)
all_usernames = df_post_comments['username']

# Get a cool name for each user and store the mapping in the database
cooler_names = {username: generate_slug(3) for index, username in enumerate(all_usernames)}
df_username_mapping = pd.DataFrame(cooler_names.items(), columns=['username', 'cooler_name'])
df_username_mapping.to_sql("username_mapping", con, index=False)


# Mapping usernames 
df_post_comments['username'] = df_post_comments['username'].map(cooler_names)
df_post_comments.to_sql("comments", con, if_exists="replace", index=False)

## 3. Collect user data from comments

# Instanciate Instaloader
L = Instaloader()

def fetch_user_data (username):
    '''Function to fetch an Instagram user's public data.

    Parameter: 
        username str: username of the user to collect the 
        data from.
    '''

    try:
        profile = Profile.from_username(L.context, username)
    except ProfileNotExistsException:
        return f"{username} does not exists anymore"
        
    data = profile.__dict__
    del data["_context"]
    json_object = json.dumps(data, indent = 2)   
    with open(f"data/users_json/{username}_user_profile_data.json", 'w') as file_object:  
        json.dump(json_object, file_object) 

def convert_json (username):
    # Open json 
    try:
        json_file = open(f"data/users_json/{username}_user_profile_data.json")
        data = json.loads(json.load(json_file))["_node"]
    except json.decoder.JSONDecodeError:
        return

    # Useless json keys
    ban = ['edge_felix_video_timeline', 'edge_owner_to_timeline_media', 'edge_saved_media', 
            'edge_media_collections', 'edge_related_profiles']

    # Defining basic keys. Those are name, follower count, bio, etc. Basic infos
    basic_keys = [key for key in data.keys() if key not in ban]
    basic_info = flatten({key:data[key] for key in basic_keys})
    df_current_user = pd.DataFrame([basic_info])

    # Getting the data for all posts that are contained in lists. 
    df_current_user["video_count"] = data['edge_felix_video_timeline']["count"]
    df_current_user["post_count"] = data['edge_owner_to_timeline_media']["count"]

    last_12_posts = dict()
    posts = data['edge_owner_to_timeline_media']["edges"]
    last_12_posts["username"] = data["username"]
    last_12_posts["video_views"] = [post["node"]["video_view_count"] if "video_view_count" in post["node"].keys() else np.nan for post in posts]
    last_12_posts["display_url"] = [post["node"]["display_url"] for post in posts]
    last_12_posts["thumbnail_src"] = [post["node"]["thumbnail_src"] for post in posts]
    last_12_posts["accessibility_caption"] = [post["node"]["accessibility_caption"] for post in posts]
    last_12_posts["is_video"] = [post["node"]["is_video"] for post in posts]
    last_12_posts["likes"] = [post["node"]["edge_liked_by"]["count"] for post in posts]
    last_12_posts["comments"] = [post["node"]["edge_media_to_comment"]["count"] for post in posts]
    last_12_posts["timestamp"] = [post["node"]["taken_at_timestamp"] for post in posts]
    df_last_12_posts = pd.DataFrame(last_12_posts)


    # Changing list type to str as sqlite3 doesn't accept this type
    list_features = ['bio_links', 'biography_with_entities_entities', 'edge_mutual_followed_by_edges', 'pronouns']
    for column in list_features: 
        if column in df_current_user.columns:
            df_current_user[column] = df_current_user[column].apply(lambda x:  "_LIST_SEPARATOR_".join(x))

    return df_current_user, df_last_12_posts

# Query post_id not in comments table 
query = """
SELECT  
    DISTINCT
    username
FROM comments 
WHERE 1=1
    AND username NOT IN (SELECT username FROM users)
"""

# Loading usernames to scrape from db
usernames_to_scrape = pd.read_sql_query(query, con)["username"]
print(f"Usernames to scrape: {len(usernames_to_scrape)}")


for index, username in enumerate(tqdm(usernames_to_scrape[:5])): 
    fetch_user_data(username)
    
    try:
        df_current_user, df_current_last_12_posts = convert_json(username)
    except FileNotFoundError: # If an error happened while fetching the data, no file
        continue

    # Not all users have the same numbaer of columns returned and SQL needs same cols to use 'append'
    df_users = pd.read_sql_query('select * from users', con)
    df_users = pd.concat([df_users, df_current_user]).drop_duplicates()
    df_users.to_sql("users", con, if_exists="replace", index=False)
    df_current_last_12_posts.to_sql("last_12_posts", con, if_exists="append", index=False)

    # Profile pic
    for username, profile_pic_url in df_current_user[["username",  "profile_pic_url"]].values: 
        try:
            urllib.request.urlretrieve(profile_pic_url, f"data/photos/user_profile_pictures/{username}_pp_user_photo.png")
        except Exception as e:
            print(username, e, end='\r')

    # Last 12 posts
    df_current_last_12_posts = df_current_last_12_posts.reset_index()
    for username, display_url, index in df_current_last_12_posts[["username", "display_url", "index"]].values: 
        urllib.request.urlretrieve(display_url, f"data/photos/user_last_12_posts/{username}_{index}_user_photo.png") 



### 3.4 Screenshot bio url landing page

# Query users having a link and remove those with NA
query = """
SELECT 
    DISTINCT 
    username
    , external_url 
FROM users 
WHERE external_url IS NOT NULL"""

df_user_urls = pd.read_sql_query(query, con)
df_user_urls

browser = Browser("chrome")
for username, external_url in tqdm(df_user_urls.values):
    try:
        browser.visit(external_url)
        browser.driver.save_screenshot(f"data/url_screenshot/{username}_external_url_screenshot.png")
        browser = mw.launch_driver("/Users/marclamy/Desktop/main file/code/igbot_final/chromedriver")
    except:
        ...

## 4. Giving cooler names to the users 

from coolname import generate_slug

# Loading all dfs and creating a list of it
df_comments = pd.read_sql_query('select * from comments', con)
df_users = pd.read_sql_query('select * from users', con)
df_last_12_posts = pd.read_sql_query('select * from last_12_posts', con)
all_dfs = [df_comments, df_users, df_last_12_posts]

# List all usernames
all_usernames = [username for df in all_dfs for username in df['username']]
print(len(all_usernames))
all_usernames = list(set(all_usernames))
print(len(all_usernames))

pd.read_sql_query('select * from comments', con)[['username']]['username'].map()

df_users['username'].head(10), df_comments['username'].head(10), df_last_12_posts['username'].head(10)

df_last_12_posts
